+++ Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("NiumaClaw.Team")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.0.16.0")]
-[assembly: AssemblyInformationalVersion("1.0.16+agent-running-ui.20260701")]
+[assembly: AssemblyFileVersion("1.0.17.0")]
+[assembly: AssemblyInformationalVersion("1.0.17+static-generated-downloads.20260702")]
 [assembly: AssemblyProduct("NiumaClaw.Team")]
 [assembly: AssemblyTitle("NiumaClaw.Team")]
-[assembly: AssemblyVersion("1.0.16.0")]
+[assembly: AssemblyVersion("1.0.17.0")]